+++ packages/gatsby-plugin-sharp/src/__tests__/index.js
@@ -1,5 +1,4 @@
 const path = require(`path`)
-const sharp = require(`sharp`)
 const fs = require(`fs-extra`)
 jest.mock(`../scheduler`)
 
@@ -16,6 +15,7 @@ jest.mock(`gatsby/dist/redux/actions`, () => {
   }
 })
 
+const sharp = require(`sharp`)
 const { scheduleJob } = require(`../scheduler`)
 scheduleJob.mockReturnValue(Promise.resolve())
 fs.ensureDirSync = jest.fn()
@@ -23,6 +23,7 @@ fs.existsSync = jest.fn().mockReturnValue(false)
 
 const {
   base64,
+  generateBase64,
   fluid,
   fixed,
   queueImageResizing,
@@ -30,6 +31,11 @@ const {
   stats,
   setBoundActionCreators,
 } = require(`../`)
+
+const {
+  getPluginOptionsDefaults,
+  setPluginOptions,
+} = require(`../plugin-options`)
 
 jest.mock(`gatsby-cli/lib/reporter`, () => {
   return {
@@ -483,6 +489,120 @@ describe(`gatsby-plugin-sharp`, () => {
 
       expect(result).toMatchSnapshot()
     })
+
+    it(`should cache same image`, async () => {
+      const file1 = getFileObject(absolutePath)
+      const file2 = getFileObject(absolutePath)
+      const file3 = getFileObject(absolutePath, `test`, `new-image`)
+      // change file of file3
+      file3.base = path.join(__dirname, `images/144-density.png`)
+
+      const result = await base64({
+        file: file1,
+        args,
+      })
+      const result2 = await base64({
+        file: file2,
+        args,
+      })
+      const result3 = await base64({
+        file: file3,
+        args,
+      })
+
+      // I would like to test sharp being executed but I don't really know how to mock that beast :p
+      expect(result).toEqual(result2)
+      expect(result).not.toEqual(result3)
+    })
+
+    // Matches base64 string in snapshot, converts to jpg to force use of bg
+    // Testing pixel colour for a match would be better
+    it(`should support option: 'background'`, async () => {
+      const result = await generateBase64({
+        file: getFileObject(path.join(__dirname, `images/alphatest.png`)),
+        args: {
+          background: `#ff0000`,
+          toFormatBase64: `jpg`,
+        },
+      })
+
+      expect(result).toMatchSnapshot()
+    })
+
+    describe(`should support option: 'base64Width'`, () => {
+      // Uses 'generateBase64()` directly to avoid `base64()` caching affecting results.
+      it(`should support a configurable width`, async () => {
+        const result = await generateBase64({
+          file,
+          args: { base64Width: 42 },
+        })
+
+        expect(result.width).toEqual(42)
+      })
+
+      it(`should support a configurable default width`, async () => {
+        setPluginOptions({ base64Width: 32 })
+
+        const result = await generateBase64({
+          file,
+          args,
+        })
+
+        expect(result.width).toEqual(32)
+        setPluginOptions(getPluginOptionsDefaults())
+      })
+
+      it(`width via arg overrides global default`, async () => {
+        setPluginOptions({ base64Width: 32 })
+
+        const result = await generateBase64({
+          file,
+          args: { base64Width: 42 },
+        })
+
+        expect(result.width).toEqual(42)
+        setPluginOptions(getPluginOptionsDefaults())
+      })
+    })
+
+    describe(`should support options: 'toFormatBase64' and 'forceBase64Format'`, () => {
+      it(`should support a different image format for base64`, async () => {
+        const result = await generateBase64({
+          file,
+          args: { toFormatBase64: `webp` },
+        })
+
+        expect(result.src).toEqual(
+          expect.stringMatching(/^data:image\/webp;base64/)
+        )
+      })
+
+      it(`should support a configurable default base64 image format`, async () => {
+        setPluginOptions({ forceBase64Format: `webp` })
+        const result = await generateBase64({
+          file,
+          args,
+        })
+
+        expect(result.src).toEqual(
+          expect.stringMatching(/^data:image\/webp;base64/)
+        )
+        setPluginOptions(getPluginOptionsDefaults())
+      })
+
+      it(`image format via arg overrides global default`, async () => {
+        setPluginOptions({ forceBase64Format: `png` })
+        const result = await generateBase64({
+          file,
+          args: { toFormatBase64: `webp` },
+        })
+
+        expect(result.src).toEqual(
+          expect.stringMatching(/^data:image\/webp;base64/)
+        )
+        setPluginOptions(getPluginOptionsDefaults())
+      })
+    })
   })
 
   describe(`image quirks`, () => {
@@ -578,7 +698,7 @@ describe(`gatsby-plugin-sharp`, () => {
   })
 })
 
-function getFileObject(absolutePath, name = `test`) {
+function getFileObject(absolutePath, name = `test`, contentDigest = `1234`) {
   const parsedPath = path.parse(absolutePath)
   return {
     id: `${absolutePath} absPath of file`,
@@ -587,7 +707,7 @@ function getFileObject(absolutePath, name = `test`) {
     absolutePath,
     extension: `png`,
     internal: {
-      contentDigest: `1234`,
+      contentDigest,
     },
   }
 }