+++ packages/gatsby-recipes/src/providers/utils/get-graphql-fields.js
@@ -1,16 +1,18 @@
-const {
-  makeExecutableSchema,
-  SchemaDirectiveVisitor,
-} = require(`graphql-tools`)
+const { makeExecutableSchema } = require(`@graphql-tools/schema`)
+const { SchemaDirectiveVisitor } = require(`@graphql-tools/utils`)
 
-const gqlFieldsToObject = fields =>
+const gqlFieldsToArray = fields =>
   Object.entries(fields).reduce((acc, [key, value]) => {
-    acc[key] = {
+    const metadata = value.metadata || {}
+    const field = {
+      id: key,
       type: value.type,
-      metadata: value.metadata,
+      name: key,
+      ...metadata,
     }
-    return acc
-  }, {})
+
+    return [...acc, field]
+  }, [])
 
 class MetadataDirective extends SchemaDirectiveVisitor {
   visitFieldDefinition(field) {
@@ -55,7 +57,7 @@ module.exports = (typeDefs, { metadata } = {}) => {
     .map(([key, value]) => {
       return {
         name: key,
-        fields: gqlFieldsToObject(value._fields),
+        fields: gqlFieldsToArray(value._fields),
       }
     })
 }