+++ www/src/views/starter-library/starter-list.js
@@ -12,7 +12,7 @@ import {
 } from "../shared/styles"
 import ThumbnailLink from "../shared/thumbnail"
 import EmptyGridItems from "../shared/empty-grid-items"
-import V2Icon from "!raw-loader!../../assets/icons/v2icon.svg"
+import { V2Icon } from "../../assets/icons"
 import { get } from "lodash-es"
 
 const StartersList = ({ urlState, starters, count }) => {
@@ -124,14 +124,15 @@ const StartersList = ({ urlState, starters, count }) => {
                             }}
                           >
                             <span
-                              dangerouslySetInnerHTML={{ __html: V2Icon }}
                               sx={{
                                 color: `textMuted`,
                                 mb: 0,
                                 mr: 2,
                                 "& svg": { height: 12, width: 12 },
                               }}
-                            />
+                            >
+                              <V2Icon />
+                            </span>
                             {` `}
                             v2
                           </span>