+++ www/src/views/starter/source.js
@@ -10,7 +10,8 @@ const Source = ({ startersYaml, repoUrl, starter }) => (
   <div
     sx={{
       display: `flex`,
-      borderTop: t => `1px solid ${t.colors.ui.border}`,
+      borderTop: 1,
+      borderColor: `ui.border`,
       fontFamily: `heading`,
       mx: 6,
       [mediaQueries.sm]: { borderTop: 0 },
@@ -72,7 +73,8 @@ const Source = ({ startersYaml, repoUrl, starter }) => (
         sx={{
           "&&": {
             borderBottom: 0,
-            marginRight: t => t.space[4],
+            mr: 4,
+            color: `link.color`,
           },
         }}
       >
@@ -85,6 +87,7 @@ const Source = ({ startersYaml, repoUrl, starter }) => (
         sx={{
           "&&": {
             borderBottom: 0,
+            color: `link.color`,
           },
         }}
       >