+++ packages/gatsby-image/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.16](https://github.com/gatsbyjs/gatsby/compare/gatsby-image@2.4.15...gatsby-image@2.4.16) (2020-08-12)
+
+### Bug Fixes
+
+- **gatsby-image:** Fix typings for fixed and fluid props ([#24767](https://github.com/gatsbyjs/gatsby/issues/24767)) ([08e0aa1](https://github.com/gatsbyjs/gatsby/commit/08e0aa1))
+
+## [2.4.15](https://github.com/gatsbyjs/gatsby/compare/gatsby-image@2.4.14...gatsby-image@2.4.15) (2020-08-10)
+
+**Note:** Version bump only for package gatsby-image
+
+## [2.4.14](https://github.com/gatsbyjs/gatsby/compare/gatsby-image@2.4.13...gatsby-image@2.4.14) (2020-07-29)
+
+### Bug Fixes
+
+- **gatsby-image:** add required check for one of fluid or fixed ([#25371](https://github.com/gatsbyjs/gatsby/issues/25371)) ([ce7ba8a](https://github.com/gatsbyjs/gatsby/commit/ce7ba8a))
+
+## [2.4.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-image@2.4.12...gatsby-image@2.4.13) (2020-07-09)
+
+**Note:** Version bump only for package gatsby-image
+
+## [2.4.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-image@2.4.11...gatsby-image@2.4.12) (2020-07-02)
+
+**Note:** Version bump only for package gatsby-image
+
+## [2.4.11](https://github.com/gatsbyjs/gatsby/compare/gatsby-image@2.4.10...gatsby-image@2.4.11) (2020-07-01)
+
+**Note:** Version bump only for package gatsby-image
+
+## [2.4.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-image@2.4.9...gatsby-image@2.4.10) (2020-07-01)
+
+**Note:** Version bump only for package gatsby-image
+
+## [2.4.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-image@2.4.8...gatsby-image@2.4.9) (2020-06-24)
+
+**Note:** Version bump only for package gatsby-image
+
+## [2.4.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-image@2.4.7...gatsby-image@2.4.8) (2020-06-22)
+
+### Bug Fixes
+
+- **docs:** change bash to shell in code language blocks ([#22899](https://github.com/gatsbyjs/gatsby/issues/22899)) ([6b6b2f2](https://github.com/gatsbyjs/gatsby/commit/6b6b2f2))
+
 ## [2.4.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-image@2.4.6...gatsby-image@2.4.7) (2020-06-09)
 
 **Note:** Version bump only for package gatsby-image