+++ packages/gatsby-plugin-sitemap/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.11](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sitemap@2.4.10...gatsby-plugin-sitemap@2.4.11) (2020-07-09)
+
+**Note:** Version bump only for package gatsby-plugin-sitemap
+
+## [2.4.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sitemap@2.4.9...gatsby-plugin-sitemap@2.4.10) (2020-07-02)
+
+**Note:** Version bump only for package gatsby-plugin-sitemap
+
+## [2.4.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sitemap@2.4.8...gatsby-plugin-sitemap@2.4.9) (2020-07-01)
+
+**Note:** Version bump only for package gatsby-plugin-sitemap
+
+## [2.4.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sitemap@2.4.7...gatsby-plugin-sitemap@2.4.8) (2020-07-01)
+
+**Note:** Version bump only for package gatsby-plugin-sitemap
+
+## [2.4.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sitemap@2.4.6...gatsby-plugin-sitemap@2.4.7) (2020-06-24)
+
+**Note:** Version bump only for package gatsby-plugin-sitemap
+
+## [2.4.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sitemap@2.4.5...gatsby-plugin-sitemap@2.4.6) (2020-06-22)
+
+**Note:** Version bump only for package gatsby-plugin-sitemap
+
 ## [2.4.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sitemap@2.4.4...gatsby-plugin-sitemap@2.4.5) (2020-06-09)
 
 **Note:** Version bump only for package gatsby-plugin-sitemap